testing moved and builds. doesn't run in tests
