+++ libs/designsystem/src/lib/index.ts
@@ -5,10 +5,10 @@ export * from '@kirbydesign/designsystem/helpers';
 export * from '@kirbydesign/designsystem/types';
 export * from '@kirbydesign/designsystem/shared';
 export * from './scss/scss-helper';
-export * from './testing';
 
 export * from './components/charts/chart-config';
 export * from '@kirbydesign/designsystem/card';
+export * from '@kirbydesign/designsystem/testing';
 
 export { KirbyModule } from './kirby.module';
 export { KirbyExperimentalModule } from './kirby-experimental.module';
