+++ libs/designsystem/testing/src/element-css-custom-matchers.ts
@@ -1,6 +1,6 @@
 import { ColorHelper, ThemeColorDefinition } from '@kirbydesign/core';
 import jasmine from 'jasmine-core';
-import { TestHelper } from 'src/lib/testing';
+import { TestHelper } from './test-helper';
 
 import CustomMatcherFactoriess = jasmine.CustomMatcherFactories;
 import CustomMatcherResult = jasmine.CustomMatcherResult;
