+++ libs/designsystem/tsconfig.spec.json
@@ -12,6 +12,8 @@
       "@kirbydesign/designsystem/shared": ["shared/index.ts"],
       "@kirbydesign/designsystem/card": ["card/index.ts"],
       "@kirbydesign/designsystem/types": ["types/index.ts"],
+      "@kirbydesign/designsystem/testing":["testing/index.ts"],
+
     }
   },
   "files": ["src/test.ts"],
